[step4] refactor: 미사용 코드 제거 및 테스트 이동
