+++ README.md
@@ -60,3 +60,13 @@
   - 지뢰가 없는 셀을 열면 셀이 열리고 인접한 빈 셀도 열린다.
   - 지뢰가 있는 셀을 열면 게임이 종료된다.
   - 모든 지뢰가 아닌 셀이 열리면 승리 조건이 충족된다.
+## step4
+### 기능 목록
+- Directions
+  - 일치하는 셀의 정확한 개수를 반환한다.
+  - 일치하는 셀이 없을 때 0을 반환한다.
+  - 그리드의 경계에서의 엣지 케이스를 처리한다.
+  - 그리드의 모서리에서의 엣지 케이스를 처리한다.
+- MineField
+  - 지뢰 주변의 빈 셀들에 올바른 숫자 힌트가 추가된다.
+  - 숫자 힌트가 추가되어도 지뢰의 위치는 변하지 않는다.
